basic tests run again
